packages/api/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis[hiredis]>=5.0.0
pydantic-settings>=2.0.0
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1          # passlib incompatible with bcrypt>=4.1
python-multipart>=0.0.9
httpx>=0.27.0
anthropic>=0.27.0
structlog>=24.0.0
python-dotenv>=1.0.0
cryptography>=42.0.0
Pillow>=10.0.0

# ── Dev / Test ─────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0
ruff>=0.4.0
